+++ pages/pointLock.html
@@ -1,28 +1,26 @@
 <!DOCTYPE html>
 <html>
-
 <head>
-    <title>PointLock Test</title>
-    <meta name="viewport" content="width=device-width, user-scalable=no, minimum-scale=1.0, maximum-scale=1.0">
+  <title>PointLock Test</title>
+  <meta name="viewport" content="width=device-width, user-scalable=no, minimum-scale=1.0, maximum-scale=1.0">
     <style>
-        html,
-        body {
+        html, body 
+        {
             margin: 0;
             height: 100%;
         }
-
-        #c {
+        #c 
+        {
             width: 100%;
             height: 100%;
             display: block;
         }
-
         #blocker {
             display: block;
             position: absolute;
             width: 100%;
             height: 100%;
-            background-color: rgba(0, 0, 0, 0.5);
+            background-color: rgba(0,0,0,0.5);
         }
 
         #instructions {
@@ -39,13 +37,148 @@
             cursor: pointer;
         }
 
-        .ui-overlay {
-            display: block !important;
-            opacity: 1 !important;
+        #interactionBlocker {
+            display: none;
+            position: absolute;
+            width: 100%;
+            height: 100%;
+        }
+
+        #interactDesc {
+            width: 100%;
+            height: 100%;
+
+            display: none;
+            flex-direction: column;
+            justify-content: center;
+            align-items: center;
+
+            text-align: center;
+            font-size: 14px;
+            cursor: pointer;
+        }
+
+        /* 
+            I wont lie, I was lazy and didnt want to do the styling for the pop up.
+            So I asked claude to design a responsive popup styles for me,
+            and I think he did quite a good job.
+        */
+        /* The overlay - covers the entire screen and darkens the background */
+        .popup-overlay {
+            display: none; /* Hidden by default */
+            position: fixed; /* Stays in place when scrolling */
+            top: 0;
+            left: 0;
+            width: 100%;
+            height: 100%;
+            background-color: rgba(0, 0, 0, 0.5); /* Semi-transparent black */
+            z-index: 1000; /* Ensures it appears above other content */
+        }
+
+        /* The actual popup container */
+        .popup-container {
+            position: absolute;
+            top: 50%;
+            left: 50%;
+            transform: translate(-50%, -50%); /* Centers the popup perfectly */
+            background-color: white;
+            border-radius: 8px;
+            box-shadow: 0 4px 20px rgba(0, 0, 0, 0.3);
+            max-width: 90%;
+            max-height: 90%;
+            overflow: auto; /* Allows scrolling if content is too large */
+            min-width: 300px;
+        }
+
+        /* Header section with title and close button */
+        .popup-header {
+            display: flex;
+            justify-content: space-between;
+            align-items: center;
+            padding: 20px;
+            border-bottom: 1px solid #eee;
+        }
+
+        .popup-title {
+            margin: 0;
+            font-size: 20px;
+            color: #333;
+        }
+
+        /* Close button styling */
+        .popup-close {
+            background: none;
+            border: none;
+            font-size: 24px;
+            cursor: pointer;
+            color: #666;
+            padding: 0;
+            width: 30px;
+            height: 30px;
+            display: flex;
+            align-items: center;
+            justify-content: center;
+            border-radius: 50%;
+        }
+
+        .popup-close:hover {
+            background-color: #f0f0f0;
+            color: #333;
+        }
+
+        /* Content area where text and images are displayed */
+        .popup-content {
+            padding: 20px;
+            line-height: 1.6;
+        }
+
+        /* Styling for images within the popup */
+        .popup-content img {
+            max-width: 100%;
+            height: auto;
+            border-radius: 4px;
+            margin: 10px 0;
+            display: block;
+        }
+
+        /* Text styling within the popup */
+        .popup-content p {
+            margin: 0 0 15px 0;
+            color: #333;
+        }
+
+        .popup-content h1, .popup-content h2, .popup-content h3 {
+            color: #333;
+            margin-top: 0;
+        }
+
+        /* Animation for smooth popup appearance */
+        .popup-overlay.show {
+            display: block;
+            animation: fadeIn 0.3s ease-out;
+        }
+
+        .popup-container.show {
+            animation: slideIn 0.3s ease-out;
+        }
+
+        @keyframes fadeIn {
+            from { opacity: 0; }
+            to { opacity: 1; }
+        }
+
+        @keyframes slideIn {
+            from { 
+                opacity: 0;
+                transform: translate(-50%, -50%) scale(0.7);
+            }
+            to { 
+                opacity: 1;
+                transform: translate(-50%, -50%) scale(1);
+            }
         }
     </style>
 </head>
-
 <body>
     <div id="blocker">
         <div id="instructions">
@@ -53,9 +186,28 @@
                 Click to play
             </p>
             <p>
-                Move: WASD or Arrow Keys<br />
-                Look: MOUSE
+                Move: WASD or Arrow Keys<br/>
+                Look: MOUSE or Q & E (Only When in GUI Mode) <br>
+                Activate/Deactivate GUI Mode: Middle Mouse button
             </p>
+        </div>
+    </div>
+    <div id="popup" class="popup-overlay">
+        <div class="popup-container">
+            <div class="popup-header">
+                <h2 id="popup-title" class="popup-title">Popup Title</h2>
+                <button class="popup-close" id="popup-close-btn">&times;</button>
+            </div>
+            <div id="popup-content" class="popup-content">
+            </div>
+        </div>
+    </div>
+    <div id="interactionBlocker">
+        <div id="interactDesc">
+            <h2>
+                Press F <br/>
+                To Interact
+            </h2>
         </div>
     </div>
     <canvas id="c"></canvas>
@@ -67,8 +219,7 @@
             }
         }
     </script>
-    <script src="../public/js/pointerLockTest.js" type="module">
+    <script src="../public/js/popUpsTest.js" type="module">
     </script>
 </body>
-
 </html>
