+++ public/js/pointerLockTest.js
@@ -147,6 +147,7 @@ class MinMaxGUIHelper {
     }
 }
 
+//This is the color gui for the directional light
 class ColorGUIHelper {
     constructor(object, prop) {
         this.object = object;
@@ -160,6 +161,7 @@ class ColorGUIHelper {
     }
 }
 
+//This is the gui for fog to control legit everything
 class FogGUIHelper {
     constructor(fog, camera) {
         this.fog = fog;
@@ -193,6 +195,126 @@ class FogGUIHelper {
     }
 }
 
+//The object for the popUp mainly its detection circle
+class popUpCircle {
+    constructor(posX, posY, posZ, radius) {
+
+        this.cameraInside = false;
+
+        this.position = new THREE.Vector3(posX, posY, posZ);
+
+        this.geometryRadius = radius;
+
+        this.circleObject = null;
+    }
+
+    createSphereRadius(scene) {
+        if(this.circleObject) scene.remove(this.circleObject);
+
+        const newSphereMaterial = new THREE.MeshBasicMaterial({
+            color: 0x00C6FF,
+            opacity: 0.5,
+            transparent: true,
+            wireframe: true,
+        });
+
+        const geometry = new THREE.CircleGeometry(this.geometryRadius);
+
+        this.circleObject = new THREE.Mesh(geometry, newSphereMaterial);
+        this.circleObject.position.set(this.position.x, this.position.y, this.position.z);
+        this.circleObject.rotation.x = (90 * Math.PI) / 180;
+
+        scene.add(this.circleObject);
+        return(this.circleObject);
+    }
+
+    checkForIntersection(camera) {
+        const cameraPosition = new THREE.Vector3();
+        camera.getWorldPosition(cameraPosition);
+
+        const sphereCenter = this.circleObject.position.clone();
+
+        const distance = cameraPosition.distanceTo(sphereCenter);
+
+        if (distance < this.geometryRadius) {
+            this.cameraInside = true;
+        } else if ( distance >= this.geometryRadius) {
+            this.cameraInside = false;
+        }
+    }
+}
+
+// Pop Up Object functionality
+const PopupManager = {
+
+    overlay: document.getElementById('popup'),
+    container: document.querySelector('.popup-container'),
+    title: document.getElementById('popup-title'),
+    content: document.getElementById('popup-content'),
+
+    popUpActive: false,
+
+    // This just unhides the pop up nothing special
+    show: function(title = 'Popup', content = '') {
+
+        this.title.textContent = title;
+
+        this.content.innerHTML = content;
+
+        this.overlay.style.display = 'block';
+        this.overlay.classList.add('show');
+        this.container.classList.add('show');
+
+        document.body.style.overflow = 'hidden';
+        document.body.style.cursor = 'default';
+    },
+
+    // self explanatory...
+    hide: function() {
+
+        this.overlay.classList.remove('show');
+        this.container.classList.remove('show');
+
+        setTimeout(() => {
+            this.overlay.style.display = 'none';
+        }, 300);
+
+        document.body.style.overflow = 'auto';
+
+        document.body.style.cursor = 'none';
+    },
+
+    // This basically is where it defines if it is a image or text pop up. Can be edited to do anything tbh
+    generatePopup: function(title, config) {
+        let content = '';
+
+        if (config.text) {
+            content += `<p>${config.text}</p>`;
+        }
+
+        if (config.image) {
+            content += `<img src="${config.image.src}" alt="${config.image.alt || ''}">`;
+            if (config.image.caption) {
+                content += `<p><em>${config.image.caption}</em></p>`;
+            }
+        }
+
+        if (config.html) {
+            content += config.html;
+        }
+        
+        this.show(title, content);
+    }
+};
+
+function closePopup() {
+    setTimeout(() => {
+        PopupManager.hide();
+    }, 300);
+    PopupManager.popUpActive = false;
+}
+
+//This is the barrier box object
 class boundaryBox {
     constructor(minX, maxX, minY, maxY, minZ, maxZ) {
         this.min = new THREE.Vector3(minX, minY, minZ);
@@ -425,6 +547,56 @@ function setupCameraBoundaries(scene, camera, controls) {
   return boundary;
 }
 
+const informationArray = [
+    // This is the sample for the popUps, you can use these however you want. 
+    // I am hoping maybe in the future I could setup like a gallery type popup
+    // that has multiple images that you can scroll through or maybe
+    // a fully custom popup that can be designed anyway you want 
+    // text: '',
+    // image: {
+    //     src: '',
+    //     alt: '',
+    //     caption: ''
+    // },
+    // html: 
+    {
+        text: 'Hopefully I can get this to work',
+        image: {
+            src: 'https://picsum.photos/350/200',
+            alt: 'Filler Image',
+            caption: 'Cool images'
+        },
+        html: '<p><strong>This is a test!</strong></p>'
+    },
+    {
+        text: 'Bill\'s Cleaners is cool',
+        image: {
+            src: 'https://picsum.photos/350/200',
+            alt: 'Filler Image',
+            caption: 'Cool images'
+        },
+        html: '<p><strong>Heard he was good at bowling!</strong></p>'
+    },
+    {
+        text: 'Pool, Beer, and Dominos!',
+        image: {
+            src: 'https://picsum.photos/350/200',
+            alt: 'Filler Image',
+            caption: 'Cool images'
+        },
+        html: '<p><strong>The fan was a pain I heard</strong></p>'
+    },
+    {
+        text: 'Records!',
+        image: {
+            src: 'https://picsum.photos/350/200',
+            alt: 'Filler Image',
+            caption: 'Cool images'
+        },
+        html: '<p><strong>Lots of goooood music here</strong></p>'
+    }
+]
+
 function main() {
     // Set up the custom shader chunks for the advanced fog effect
     THREE.ShaderChunk.fog_fragment = `
@@ -461,20 +633,20 @@ function main() {
       #endif
     #endif`;
     
-    //CHATGPT SAVING THE DAY!!!!!
-    //I couldnt figure out what the problem was it kept saying shader issue
-    //so I looked and looked and looked through my code and the shaders were being properly
-    //loaded and setup so I was so confused what the problem was. 
-    //so I finally decided im going to ask AI. Claude couldnt figure it out,
-    //but my good old homie ChatGPT somehow figured out that i needed to initialize the worldPosition variable in
-    //the fog vertex using vec4. I was shocked and honestly relieved. 
+    // CHATGPT SAVING THE DAY!!!!!
+    // I couldnt figure out what the problem was it kept saying shader issue
+    // so I looked and looked and looked through my code and the shaders were being properly
+    // loaded and setup so I was so confused what the problem was. 
+    // so I finally decided im going to ask AI. Claude couldnt figure it out,
+    // but my good old homie ChatGPT somehow figured out that i needed to initialize the worldPosition variable in
+    // the fog vertex using vec4. I was shocked and honestly relieved. 
 
-    //Prompt to AI -
-    //I am currently trying to implement a split view camera into my program as a 
+    // Prompt to AI -
+    // I am currently trying to implement a split view camera into my program as a 
     // test to configure the camera properly. Everything was working fine until I started implementing the camera code. 
-    //I am getting this error message and I need you to see if you can configure what it is meaning,
-    //also see if you can find the solution if possible.
-    //-Error message goes here but it is way to big-
+    // I am getting this error message and I need you to see if you can configure what it is meaning,
+    // also see if you can find the solution if possible.
+    // -Error message goes here but it is way to big-
     THREE.ShaderChunk.fog_vertex = `
     #ifdef USE_FOG
       vec4 worldPosition = modelMatrix * vec4(position, 1.0);
@@ -555,7 +727,7 @@ function main() {
     controls.minPolarAngle = (60 * Math.PI) / 180;
 
     instructions.addEventListener( 'click', function () {
-        if (!isGUIMode) {
+        if (!isGUIMode && !PopupManager.popUpActive) {
             controls.lock();
             instructionsActive = false;
             instructions.style.display = 'none';
@@ -563,13 +735,31 @@ function main() {
         }
     });
 
+    document.getElementById('popup').addEventListener('click', function(e) {
+        if (e.target === this) {
+            closePopup();
+            controls.lock();
+            PopupManager.popUpActive = false;
+        }
+    });
+
+    document.getElementById('popup-close-btn').addEventListener('click', function(e) {
+        if (e.target === this) {
+            closePopup();
+            controls.lock();
+            PopupManager.popUpActive = false;
+        }
+    });
+
     controls.addEventListener( 'unlock', function () {
         controls.unlock();
         console.log('Controls have been unlocked');
-        if(!isGUIMode) {
+        if(!isGUIMode && !PopupManager.popUpActive ) {
             instructionsActive = true;
             instructions.style.display = '';
             blocker.style.display = '';
+            document.getElementById('interactionBlocker').style.display = 'none';
+            document.getElementById('interactDesc').style.display = 'none';
         }
     });
 
@@ -582,13 +772,13 @@ function main() {
     function toggleGUIMode() {
         isGUIMode = !isGUIMode;
 
-        if ( isGUIMode && !instructionsActive) {
+        if ( isGUIMode && !instructionsActive && !PopupManager.popUpActive ) {
             if ( controls.isLocked ) {
                 controls.unlock();
             }
             console.log('GUI Mode: Activated - Mouse is now free for GUI interaction');
             updateGUIVisibility();
-        } else if(!instructionsActive) {
+        } else if(!instructionsActive && !PopupManager.popUpActive ) {
             controls.lock();
             console.log('GUID Mode: Deactivated - Camera Controls active');
             updateGUIVisibility();
@@ -601,6 +791,8 @@ function main() {
             if (isGUIMode) {
                 element.style.pointerEvents = 'auto';
                 element.style.opacity = '1';
+                document.getElementById('interactionBlocker').style.display = 'none';
+                document.getElementById('interactDesc').style.display = 'none';
             } else {
                 element.style.pointerEvents = 'none';
                 element.style.opacity = '0.3';
@@ -634,8 +826,10 @@ function main() {
             case 'KeyD':
                 moveRight = true;
                 break;
-            
-            
+        }
+    };
+    const rotateTheCamera = function ( event ) {
+        switch (event.code) {
             case isGUIMode && 'KeyQ':
                 cameraEuler.setFromQuaternion(camera.quaternion);
                 cameraEuler.y -= -0.01 * 0.5 * 2;
@@ -648,7 +842,7 @@ function main() {
                 camera.quaternion.setFromEuler(cameraEuler);
                 break;
         }
-    };
+    }
     const onKeyUp = function ( event ) {
 
         switch ( event.code ) {
@@ -682,6 +876,7 @@ function main() {
     };
     document.addEventListener( 'keydown', onKeyDown );
     document.addEventListener( 'keyup', onKeyUp );
+    document.addEventListener( 'keydown', rotateTheCamera);
     document.addEventListener( 'mousedown', handleMiddleClick);
 
     // Store shaders that need to be updated with fogTime
@@ -713,7 +908,7 @@ function main() {
         boundaryFolder.add(cameraBoundarySystem.max, 'x', -100, 0).name('Max X');
         boundaryFolder.add(cameraBoundarySystem.min, 'z', -150, 0).name('Min Z');
         boundaryFolder.add(cameraBoundarySystem.max, 'z', 0, 150).name('Max Z');
-        //This is what claude created to setup the rotation GUI
+        // This is what claude created to setup the rotation GUI
         // Add rotation controls using degrees (more user-friendly)
         // The key insight here is that we're directly manipulating the object's properties
         // and then calling updateRotation() whenever a value changes
@@ -741,6 +936,105 @@ function main() {
         boundaryFolder.open();
     }
 
+    // Intersection pop Circles!
+    const popCirclesGUI = gui.addFolder('Popup Circles');
+    const theaterGUI = popCirclesGUI.addFolder('Theater Circle');
+    const cleanersGUI = popCirclesGUI.addFolder('Bills Circle');
+    const dominosGUI = popCirclesGUI.addFolder('Dominos Circle');
+    const recordsGUI = popCirclesGUI.addFolder('Records Circle')
+
+    // Theater circle Intersection popup
+    const theaterSphere = new popUpCircle(-32, 31, 7, 8);
+    theaterSphere.createSphereRadius(scene);
+    theaterGUI.add(theaterSphere.position, 'x', -50, 50, 1).onChange((value) => {
+        if ( theaterSphere.circleObject) {
+            theaterSphere.circleObject.position.x = value;
+        }
+    });
+    theaterGUI.add(theaterSphere.position, 'z', -20, 20, 1).onChange((value) => {
+        if ( theaterSphere.circleObject) {
+            theaterSphere.circleObject.position.z = value;
+        }
+    });
+    // Bills Cleaners
+    const cleanersSphere = new popUpCircle(-35, 31, 32, 4);
+    cleanersSphere.createSphereRadius(scene);
+    cleanersGUI.add(cleanersSphere.position, 'x', -80, 50, 0.1).onChange((value) => {
+        if ( cleanersSphere.circleObject) {
+            cleanersSphere.circleObject.position.x = value;
+        }
+    });
+    cleanersGUI.add(cleanersSphere.position, 'z', -20, 60, 0.1).onChange((value) => {
+        if ( cleanersSphere.circleObject) {
+            cleanersSphere.circleObject.position.z = value;
+        }
+    });
+    // Dominos place with THE FAN
+    const dominosSphere = new popUpCircle(-35.2, 31, 57.8, 3);
+    dominosSphere.createSphereRadius(scene);
+    dominosGUI.add(dominosSphere.position, 'x', -80, 50, 0.1).onChange((value) => {
+        if ( dominosSphere.circleObject) {
+            dominosSphere.circleObject.position.x = value;
+        }
+    });
+    dominosGUI.add(dominosSphere.position, 'z', -20, 60, 0.1).onChange((value) => {
+        if ( dominosSphere.circleObject) {
+            dominosSphere.circleObject.position.z = value;
+        }
+    });
+    // Records Shop, good music bruh
+    const recordsSphere = new popUpCircle(-36, 31, 63, 2);
+    recordsSphere.createSphereRadius(scene);
+    recordsGUI.add(recordsSphere.position, 'x', -80, 50, 0.1).onChange((value) => {
+        if ( recordsSphere.circleObject) {
+            recordsSphere.circleObject.position.x = value;
+        }
+    });
+    recordsGUI.add(recordsSphere.position, 'z', -20, 100, 0.1).onChange((value) => {
+        if ( recordsSphere.circleObject) {
+            recordsSphere.circleObject.position.z = value;
+        }
+    });
+
+    popCirclesGUI.open();
+
+    const interactListener = function ( event ) {
+        if (isGUIMode || instructionsActive) return;
+
+        switch (event.code) {
+            case 'KeyF':
+                console.log('Interacted!');
+                if (theaterSphere.cameraInside) {
+                    PopupManager.popUpActive  = true;
+                    PopupManager.generatePopup('Theater', informationArray[0]);
+                    controls.unlock();
+                    event.preventDefault();
+                    break;
+                }else if (cleanersSphere.cameraInside) {
+                    PopupManager.popUpActive  = true;
+                    PopupManager.generatePopup('Bills Cleaners', informationArray[1]);
+                    controls.unlock();
+                    event.preventDefault();
+                    break;
+                }else if (dominosSphere.cameraInside) {
+                    PopupManager.popUpActive  = true;
+                    PopupManager.generatePopup('Pool Dominos', informationArray[2]);
+                    controls.unlock();
+                    event.preventDefault();
+                    break;
+                }else if (recordsSphere.cameraInside) {
+                    PopupManager.popUpActive  = true;
+                    PopupManager.generatePopup('Record Shop', informationArray[3]);
+                    controls.unlock();
+                    event.preventDefault();
+                    break;
+                } else {
+                    break; 
+                }  
+        }
+    }
+    document.addEventListener( 'keydown', interactListener);
+
     // Add hemisphere light
     {
         const skyColor = 0xB1E1FF;
@@ -763,6 +1057,7 @@ function main() {
         lightFolder.addColor(new ColorGUIHelper(light, 'color'), 'value').name('color');
         lightFolder.add(light, 'intensity', 0, 5, 0.01);
         lightFolder.open();
+
     }
 
     // Add skybox and GUI Controls
@@ -797,6 +1092,7 @@ function main() {
             // Now that the skybox is loaded, set up the GUI control
             setupSkyboxGUI();
         });
+
     }
 
     // Create the skybox control object with a proper property that holds the current selection
@@ -809,6 +1105,8 @@ function main() {
             if (skySphereMesh && skyBoxTextures[newTextureName]) {
                 controls.disconnect();
                 document.removeEventListener( 'keydown', onKeyDown );
+                document.removeEventListener( 'keydown', interactListener );
+                document.removeEventListener( 'keydown', rotateTheCamera );
                 document.removeEventListener( 'keyup', onKeyUp );
                 document.removeEventListener( 'mousedown', handleMiddleClick);
                 setTimeout(() => {
@@ -816,6 +1114,8 @@ function main() {
                     skySphereMesh.material.needsUpdate = true;
                 }, 100);
                 document.addEventListener( 'keydown', onKeyDown );
+                document.addEventListener( 'keydown', interactListener );
+                document.addEventListener( 'keydown', rotateTheCamera );
                 document.addEventListener( 'keyup', onKeyUp );
                 document.addEventListener( 'mousedown', handleMiddleClick);
                 controls.connect(canvas);
@@ -898,6 +1198,7 @@ function main() {
             loadingDiv.style.background = 'rgba(255,0,0,0.7)';
             console.error('Error loading model:', error);
         });
+
     }
 
     function resizeRendererToDisplaySize(renderer) {
@@ -931,7 +1232,22 @@ function main() {
 
         const pointLockTime = performance.now();
 
-        if ( controls.isLocked === true || isGUIMode ){
+        if ( controls.isLocked === true && !isGUIMode) {
+            theaterSphere.checkForIntersection(camera);
+            cleanersSphere.checkForIntersection(camera);
+            dominosSphere.checkForIntersection(camera);
+            recordsSphere.checkForIntersection(camera);
+            let theCameraInside = (theaterSphere.cameraInside || cleanersSphere.cameraInside || dominosSphere.cameraInside || recordsSphere.cameraInside) ? true : false;
+            if (theCameraInside) {
+                document.getElementById('interactionBlocker').style.display = 'block';
+                document.getElementById('interactDesc').style.display = 'flex';
+            } else if (document.getElementById('interactionBlocker').style.display === 'block' && !theCameraInside) {
+                document.getElementById('interactionBlocker').style.display = 'none';
+                document.getElementById('interactDesc').style.display = 'none';
+            }
+        }
+
+        if ( controls.isLocked === true || isGUIMode  && !PopupManager.popUpActive ){
 
             const delta = ( time - prevTime ) / 1000;
